+++ src/trex/cli/__init__.py
@@ -159,6 +159,12 @@ def add_common_arguments(parser, smartseq: bool):
         "Expected format: see documentation",
         default=None,
     )
+    filter_group.add_argument(
+        "--filter-cloneids",
+        type=Path,
+        help="Text file with cloneIDs to be ignored during the analysis. Format: One cloneID per line. "
+        "Use this to remove e.g. overrepresented cloneIDs or misalignments.",
+    )
 
     output_group = parser.add_argument_group("Output directory")